// web3-login/packages/CustomAuth/src/handlers/interfaces.ts
export type TorusGenericObject = {
  [key: string]: string;
};

export type PopupResponse = {
  hashParams: { access_token: string; id_token?: string };
  instanceParams: TorusGenericObject;
};

export interface Auth0UserInfo {
  picture: string;
  email: string;
  name: string;
  sub: string;
  nickname: string;
}

export interface ExtraParams {
  [key: string]: unknown;
}

export type WebAuthnExtraParams = {
  signature?: string;
  clientDataJSON?: string;
  authenticatorData?: string;
  publicKey?: string;
  challenge?: string;
  rpOrigin?: string;
  credId?: string;
  transports?: AuthenticatorTransport[];
};
export interface TorusVerifierResponse {
  email: string;
  name: string;
  profileImage: string;
  aggregateVerifier?: string;
  verifierId: string;
  typeOfLogin: LOGIN_TYPE;
  ref?: string;
  extraVerifierParams?: WebAuthnExtraParams;
}

export interface TorusSubVerifierInfo {
  verifier: string;
  idToken: string;
  extraVerifierParams?: WebAuthnExtraParams;
}

export interface LoginWindowResponse {
  accessToken?: string;
  idToken?: string;
  ref?: string;
  extraParams?: string;
  extraParamsPassed?: string;
  state: TorusGenericObject;
}

export interface ILoginHandler {
  clientId: string;
  nonce: string;
  finalURL: URL;
  getUserInfo(params: LoginWindowResponse): Promise<TorusVerifierResponse>;
  handleLoginWindow(params: { locationReplaceOnRedirect?: boolean; popupFeatures?: string }): Promise<LoginWindowResponse>;
}

export interface TorusKeyPub {
  pubKey?: {
    pub_key_X: string;
    pub_key_Y: string;
  };
}

export interface TorusKey extends TorusKeyPub {
  privateKey: string;
}

export interface TorusAggregateVerifierResponse {
  userInfo: (TorusVerifierResponse & LoginWindowResponse)[];
}

export interface TorusSingleVerifierResponse {
  userInfo: TorusVerifierResponse & LoginWindowResponse;
}

export type TorusLoginResponse = TorusSingleVerifierResponse & TorusKey;
// ...
export interface InitParams {
  /**
   * skips the installation / check for service worker
   * @defaultValue false
   */
  skipSw?: boolean;

  /**
   * skips the init function
   * @defaultValue false
   */
  skipInit?: boolean;

  /**
   * skips the prefetching of redirect url
   * @defaultValue false
   *
   */
  skipPrefetch?: boolean;
}
// ...
export interface Auth0ClientOptions extends BaseLoginOptions {
  /**
   * Your Auth0 account domain such as `'example.auth0.com'`,
   * `'example.eu.auth0.com'` or , `'example.mycompany.com'`
   * (when using [custom domains](https://auth0.com/docs/custom-domains))
   */
  domain?: string;
  /**
   * The Client ID found on your Application settings page
   */
  client_id?: string;
  /**
   * The default URL where Auth0 will redirect your browser to with
   * the authentication result. It must be whitelisted in
   * the "Allowed Callback URLs" field in your Auth0 Application's
   * settings. If not provided here, it should be provided in the other
   * methods that provide authentication.
   */
  redirect_uri?: string;
  /**
   * The value in seconds used to account for clock skew in JWT expirations.
   * Typically, this value is no more than a minute or two at maximum.
   * Defaults to 60s.
   */
  leeway?: number;

  /**
   * The field in jwt token which maps to verifier id
   */
  verifierIdField?: string;

  /**
   * Whether the verifier id field is case sensitive
   * @defaultValue true
   */
  isVerifierIdCaseSensitive?: boolean;

  id_token?: string;

  access_token?: string;
  /**
   * The route for user info endpoint. This will be padded to domain
   * @defaultValue userinfo
   * */
  user_info_route?: string;
}

export interface SubVerifierDetails {
  typeOfLogin: LOGIN_TYPE;
  verifier: string;
  clientId: string;
  jwtParams?: Auth0ClientOptions;
  hash?: string;
  queryParameters?: TorusGenericObject;
  customState?: TorusGenericObject;
  idToken?: string;
  accessToken?: string;
}

export interface CreateHandlerParams {
  typeOfLogin: LOGIN_TYPE;
  clientId: string;
  redirect_uri: string;
  uxMode?: UX_MODE_TYPE;
  redirectToOpener?: boolean;
  jwtParams?: Auth0ClientOptions;
  customState?: TorusGenericObject;
}

export interface RedirectResultParams {
  replaceUrl?: boolean;
  clearLoginDetails?: boolean;
}

export type SingleLoginParams = SubVerifierDetails;